finalMerge

+++ database/migrations/2023_05_17_084606_create_participants_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
             $table->string('participant_name');
-            $table->integer('participant_nim');
+            $table->bigInteger('participant_nim');
             $table->string('binusian_email');
             $table->string('instagram_account');
             $table->string('following_proof1')->nullable();
